requirements.txt
# Project Management Dashboard - Python Dependencies

# CLI Framework
typer[all]==0.9.0
rich==13.7.0

# Database Drivers
psycopg2-binary==2.9.9
clickhouse-driver==0.2.6

# Data Validation
pydantic==2.5.0
pydantic-settings==2.1.0

# Configuration
python-dotenv==1.0.0

# Dashboard
dash==2.14.0
dash-bootstrap-components==1.5.0
plotly==5.18.0

# Data Processing
pandas==2.1.4

# HTTP Server (for Dash)
gunicorn==21.2.0

# Testing
pytest==7.4.3
pytest-cov==4.1.0
